+++ webapp/src/components/centerPanelPages/pageIconSelector.tsx
