+++ webapp/src/components/centerPanelPages/pageTitle.scss
