+++ webapp/src/components/centerPanelPages/pageTitle.tsx
@@ -3,16 +3,16 @@
 import React, {useState, useCallback, useContext, useEffect} from 'react'
 import {FormattedMessage, useIntl} from 'react-intl'
 
-import {BlockIcons} from '../blockIcons'
-import {Page} from '../blocks/page'
-import {Board} from '../blocks/board'
-import mutator from '../mutator'
-import isPagesContext from '../isPages'
-import Button from '../widgets/buttons/button'
-import Editable from '../widgets/editable'
-import CompassIcon from '../widgets/icons/compassIcon'
-import {Permission} from '../constants'
-import {useHasCurrentBoardPermissions} from '../hooks/permissions'
+import {BlockIcons} from '../../blockIcons'
+import {Page} from '../../blocks/page'
+import {Board} from '../../blocks/board'
+import mutator from '../../mutator'
+import isPagesContext from '../../isPages'
+import Button from '../../widgets/buttons/button'
+import Editable from '../../widgets/editable'
+import CompassIcon from '../../widgets/icons/compassIcon'
+import {Permission} from '../../constants'
+import {useHasCurrentBoardPermissions} from '../../hooks/permissions'
 
 import PageIconSelector from './pageIconSelector'
 import './pageTitle.scss'
